+++ uvm_pygen/models/logic_schema/reference_model.py
@@ -41,7 +41,6 @@ class ReferenceModelModel(BaseModel):
     model_config = ConfigDict(frozen=True)
 
     class_name: str  # e.g. "alu_reference_model"
-    transaction_type: str  # TODO: remove since transaction type is now per-connection, not global to the RM
     strategy: ReferenceModelStrategy
     implementation: ReferenceModelImplEnum
     dpi_function: str | None = None